+++ Source/DREAM3DLib/DistributionAnalysisOps/DistributionAnalysisOps.cpp
@@ -56,7 +56,7 @@ void DistributionAnalysisOps::determinemaxandminvalues(QVector<float>& data, flo
   float value;
   min = std::numeric_limits<float>::max();
   max = std::numeric_limits<float>::min();
-  for (size_t i = 0; i < data.size(); i++)
+  for (qint32 i = 0; i < data.size(); i++)
   {
     value = data[i];
     if(value > max) { max = value; }
